requirements.txt
chromadb==0.4.22
langchain==0.2.12
langchain-community==0.2.12
langchain-huggingface==0.0.3
langchain-groq==0.0.2
sentence-transformers==2.2.2
torch==2.2.2
transformers==4.41.2
groq==0.9.0
python-dotenv==1.0.1
tf-keras